This commit adds a large amount of code from atmel start. Much of it is very boiler plate and inefficient, but it works!
This boots the micro with an internal 8MHz oscillator as the core clock.

A 50kHz PWM signal is output as well as an enable signal for the FET
controller.

This also reads voltages from both of the power monitoring ICs.

Very messy at this point, but this is a good base.